Implement a parser and formatter from scratch
Previously operate-on-number relied on `number-at-point`, meaning it
could only operate on numbers tokenized as S-expression, but the newly
implemented parser understands even more number formats, and is also
able to extract a number sequence inside of a word.

A new formatter is implemented so that `operate-on-number-at-point` can
put a result in the same format as the original number.